+++ app/src/atoms/MenuList/MenuItem.stories.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react'
+import { MenuItem } from './MenuItem'
+
+import type { Story, Meta } from '@storybook/react'
+
+export default {
+  title: 'App/Atoms/MenuItem',
+  component: MenuItem,
+} as Meta
+
+const Template: Story<React.ComponentProps<typeof MenuItem>> = args => (
+  <MenuItem {...args} />
+)
+export const Primary = Template.bind({})
+Primary.args = {
+  children: 'Example menu btn',
+}
