+++ app/src/atoms/MenuList/MenuItem.tsx
@@ -11,7 +11,7 @@ import type { PrimitiveComponent } from '@opentrons/components'
 
 type BtnComponent = PrimitiveComponent<'button'>
 
-export const OverflowMenuBtn: BtnComponent = styled(Btn)`
+export const MenuItem: BtnComponent = styled(Btn)`
   width: ${TYPOGRAPHY.overflowMenuWidth};
   text-align: ${TEXT_ALIGN_LEFT};
   font-size: ${TYPOGRAPHY.fontSizeP};
